% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tukey_tests.R
\name{HSDTable}
\alias{HSDTable}
\title{Function to automagically create tukey test tables from an anova anlysis}
\usage{
HSDTable(aov, which = NULL, metric = "AUC", type = "html",
  cap = "Tukey HSD Test", label = "", digits = 3, file = "")
}
\arguments{
\item{aov}{Anova analysis of from \code{\link[stats]{aov}}}

\item{which}{This is what you factors you want to test and can either be a single string or array of strings c('F1','F2')}

\item{metric}{The metric you want to use for evaluation}

\item{type}{Type of output you would like.  valid values are 'html' and 'latex'}

\item{label}{Latex label to give the table}

\item{digits}{Number of significant digits to include in the table}

\item{file}{The file in which to output the results ("" means it simply prints to the screen)}

\item{caption}{String containing the caption of the table}
}
\description{
Function to automagically create tukey test tables from an anova anlysis
}

