% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_results.R
\name{mapKeys}
\alias{mapKeys}
\title{Map the keys from a datarun and load them into a dataframe}
\usage{
mapKeys(zipfile, datafile,
  keysheet = "https://www.dropbox.com/s/o8x5mmhjhtolkkb/feature-ranking-key.xls?dl=1")
}
\arguments{
\item{zipfile}{The location of the zipped up output of the run (can be absolute or relative)}

\item{datafile}{The file within the zip file which conatins the results (output/output2.arff)}

\item{keysheet}{The location of the keysheet to be used for mapping (can be a url or file location)}
}
\value{
A dataframe of the results with the keys mapped to human-readable labels
}
\description{
Map the keys from a datarun and load them into a dataframe
}

