% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats_tests.R
\name{summarySE}
\alias{summarySE}
\title{Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95\%)}

\item{.drop}{should combinations of variables that do not appear in the input data be
preserved (FALSE) or dropped (TRUE, default)}
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).
}

